requirements.txt
flask==3.0.0
torch==2.1.0
torchvision==0.16.0
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
transformers==4.35.0
mediapipe==0.10.7
gunicorn==21.2.0

